refactor: Move letterscheme into settings
